apps/ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "google-genai>=1.68.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.42.0",
]
